requirements.txt
discord.py>=2.3.2
gspread==5.11.4
oauth2client==4.1.3
aiohttp==3.9.3
Flask==3.0.2